Add new web crawler with Maven dependencies
A new web crawler has been added to the project. This crawler is designed to retrieve information from specified URLs and then process the information to create a JSON object. This object is then sent to a tracker for further processing. The crawler is built with dependencies managed through Maven, including jsoup for HTML parsing and json for handling JSON data. This decision was made to ensure the efficient managing of dependencies and facilitate in easy replication of the project environment.